requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-filter==2.4.0
boto3==1.12.3